src/pyproject.toml
[tool.poetry]
name = "Fastapi Boilerplate"
version = "0.1.0"
description = "Fastapi boilerplate for fast development"
authors = ["Sugeng <me@sugeng.me>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.63.0"
tortoise-orm = "^0.17.1"
fastapi-users = { version = "^6.0.0", extras = ["\"tortoise-orm\""] }
fastapi-crudrouter = "^0.7.1"
uvicorn = "^0.13.4"
gunicorn = "^20.1.0"
uvloop = "^0.15.2"
httptools = "^0.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
coverage = { version = "^5.5", extras = ["toml"] }
black = "^21.5b0"
isort = { version = "^5.8.0", extras = ["pyproject"] }
requests = "^2.25.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
# https://github.com/psf/black
line-length = 88
target_version = ['py38']
exclude = '''
(
  /(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | htmlcov
    | venv
    | .venv
  )/
)
'''

[tool.isort]
# https://github.com/timothycrosley/isort
# https://github.com/timothycrosley/isort/wiki/isort-Settings
line_length = 88
indent = '    '
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0


[tool.coverage]
# https://github.com/nedbat/coveragepy
[tool.coverage.run]

source = ["app"]
branch = true
omit = ['']

[tool.coverage.report]
exclude_lines = [
    "# noqa",
    "raise NotImplementedError"
]
